backend/requirements.txt
llama-index==0.11.20
llama-index-llms-openai>=0.2.10
llama-index-embeddings-openai>=0.2.4
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
